requirements.txt
boto3>=1.34.0
botocore>=1.34.0
rich>=13.7.0
PyYAML>=6.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
python-docx>=1.1.0
